# Eliminate potential races in ContextWatcher, refactor