--- a/internal/ctxwatch/context_watcher.go
+++ b/internal/ctxwatch/context_watcher.go
@@ -2,6 +2,7 @@ package ctxwatch
 
 import (
 	"context"
+	"sync/atomic"
 )
 
 // ContextWatcher watches a context and performs an action when the context is canceled. It can watch one context at a
@@ -9,9 +10,8 @@ import (
 type ContextWatcher struct {
 	onCancel             func()
 	onUnwatchAfterCancel func()
-	unwatchChan          chan struct{}
-	watchInProgress      bool
-	onCancelWasCalled    bool
+	canceled             chan bool
+	watching             uint32
 }
 
 // NewContextWatcher returns a ContextWatcher. onCancel will be called when a watched context is canceled.
@@ -21,7 +21,7 @@ func NewContextWatcher(onCancel func(), onUnwatchAfterCancel func()) *ContextWat
 	cw := &ContextWatcher{
 		onCancel:             onCancel,
 		onUnwatchAfterCancel: onUnwatchAfterCancel,
-		unwatchChan:          make(chan struct{}),
+		canceled:             make(chan bool),
 	}
 
 	return cw
@@ -29,36 +29,41 @@ func NewContextWatcher(onCancel func(), onUnwatchAfterCancel func()) *ContextWat
 
 // Watch starts watching ctx. If ctx is canceled then the onCancel function passed to NewContextWatcher will be called.
 func (cw *ContextWatcher) Watch(ctx context.Context) {
-	if cw.watchInProgress {
+	if swapped := atomic.CompareAndSwapUint32(&cw.watching, 0, 1); !swapped {
 		panic("Watch already in progress")
 	}
 
-	cw.onCancelWasCalled = false
+	if ctx.Done() == nil {
+		cw.watching = 0 // Known to be not watching, so no need for an atomic store.
+		return
+	}
 
-	if ctx.Done() != nil {
-		cw.watchInProgress = true
-		go func() {
-			select {
-			case <-ctx.Done():
-				cw.onCancel()
-				cw.onCancelWasCalled = true
-				<-cw.unwatchChan
-			case <-cw.unwatchChan:
-			}
-		}()
-	} else {
-		cw.watchInProgress = false
+	go cw.watch(ctx)
+}
+
+func (cw *ContextWatcher) watch(ctx context.Context) {
+	select {
+	case <-ctx.Done():
+		watching := atomic.LoadUint32(&cw.watching) == 1
+		if watching {
+			cw.onCancel()
+		}
+		cw.canceled <- watching
+
+	case cw.canceled <- false:
 	}
 }
 
 // Unwatch stops watching the previously watched context. If the onCancel function passed to NewContextWatcher was
 // called then onUnwatchAfterCancel will also be called.
 func (cw *ContextWatcher) Unwatch() {
-	if cw.watchInProgress {
-		cw.unwatchChan <- struct{}{}
-		if cw.onCancelWasCalled {
-			cw.onUnwatchAfterCancel()
-		}
-		cw.watchInProgress = false
+	watching := atomic.CompareAndSwapUint32(&cw.watching, 1, 0)
+	if !watching {
+		return
+	}
+
+	canceled := <-cw.canceled
+	if canceled {
+		cw.onUnwatchAfterCancel()
 	}
 }