Fixes #25.

Instead of bare booleans, use an atomic value which indicates if the context is being watched. Switch `onCancelWasCalled` and `unwatchChan` to a `chan bool` which communicates if `onUnwatchAfterCancel` has been called.

Modeled on paper with a state machine to ensure that the semantics of this watcher match the original intent. I find the new code to be a little more readable.

I'm fairly confident this eliminates race potential (https://github.com/hortbot/hortbot/runs/385472775 w/ my fork). I'm not entirely certain why the existing tests didn't fail in race mode, but maybe the race required concurrent callers to `Unwatch` or similar or a specific cancellation setup.

`benchstat` of this change:

```
name                           old time/op    new time/op    delta
ContextWatcherUncancellable-4    3.73ns ± 0%   16.23ns ± 1%  +335.56%  (p=0.000 n=8+10)
ContextWatcherCancelled-4         713ns ± 1%     760ns ± 2%    +6.57%  (p=0.000 n=10+10)
ContextWatcherCancellable-4       454ns ± 1%     479ns ± 0%    +5.65%  (p=0.000 n=10+9)

name                           old alloc/op   new alloc/op   delta
ContextWatcherUncancellable-4     0.00B          0.00B           ~     (all equal)
ContextWatcherCancelled-4          176B ± 0%      176B ± 0%      ~     (all equal)
ContextWatcherCancellable-4       0.00B          0.00B           ~     (all equal)

name                           old allocs/op  new allocs/op  delta
ContextWatcherUncancellable-4      0.00           0.00           ~     (all equal)
ContextWatcherCancelled-4          3.00 ± 0%      3.00 ± 0%      ~     (all equal)
ContextWatcherCancellable-4        0.00           0.00           ~     (all equal)
```

Small overall penalty for the atomic instructions. `ContextWatcherUncancellable` is impacted the most as I preserved the behavior that `Watch` should panic if it's already watching, even if `ctx.Done() == nil`. If you don't care about panicking if the new context cannot be canceled, I'm sure I can lower this, but we're already working at incredibly small timescales.