@zikaeroh 2020-01-12
After fixing this race, I'm now seeing cases where `Watch` has been called again on something that was already watching. For example, see this build (ignore the "passing", it's `gotip` so I have it set to always succeed): https://travis-ci.com/hortbot/hortbot/builds/144044051

```
panic: Watch already in progress
goroutine 38 [running]:
github.com/jackc/pgconn/internal/ctxwatch.(*ContextWatcher).Watch(0xc000275c60, 0x1105f40, 0xc0000a3440)
	/home/travis/gopath/pkg/mod/github.com/zikaeroh/pgconn@v1.2.1-0.20200112090342-32f9642f8eb8/internal/ctxwatch/context_watcher.go:38 +0xe6
github.com/jackc/pgconn.(*PgConn).ayncClose.func1(0xc0002374a0)
	/home/travis/gopath/pkg/mod/github.com/zikaeroh/pgconn@v1.2.1-0.20200112090342-32f9642f8eb8/pgconn.go:526 +0x199
created by github.com/jackc/pgconn.(*PgConn).ayncClose
	/home/travis/gopath/pkg/mod/github.com/zikaeroh/pgconn@v1.2.1-0.20200112090342-32f9642f8eb8/pgconn.go:518 +0x96
```

The code in question is:

```go
// ayncClose marks the connection as closed and asynchronously sends a cancel query message and closes the underlying
// connection.
func (pgConn *PgConn) ayncClose() {
	if pgConn.status == connStatusClosed {
		return
	}
	pgConn.status = connStatusClosed

	go func() {
		defer pgConn.conn.Close()

		ctx, cancel := context.WithTimeout(context.Background(), time.Second*15)
		defer cancel()

		pgConn.CancelRequest(ctx)

		pgConn.contextWatcher.Watch(ctx)
		defer pgConn.contextWatcher.Unwatch()

		pgConn.conn.Write([]byte{'X', 0, 0, 0, 4})
		pgConn.conn.Read(make([]byte, 1))
	}()
}
```

This appears to be an incorrect usage of `ContextWatcher`, as it's creating a fresh context and attempting to use it with `pgConn.contextWatcher`, which would likely have already been used.