@zikaeroh 2020-01-12
Per #25, this watcher was never intended to be currently accessed anyway. Closing, since it was the use that needed to be removed.